requirements.txt
# Core Application Dependencies
openai>=1.40.0
langchain-openai>=0.1.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.5.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
aiohttp>=3.9.0
loguru==0.7.3

# Testing Dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
requests-mock>=1.11.0
httpx>=0.25.0